// ClientApp/src/app/shared/models/book.ts
export class Book {
  id: number;
  title: string;
  mediaType: number;
  language: number;
  rating: number;
  imageUrl: string;
  price: number;
  authorName: string;
  categoryName: string;
  publisherName: string;
  mediaTypeName: string;
  languageName: string;
  publishedDate: Date;
  totalPages: number;
  introduction: string;
  categoryId: number;
  publisherId: number;
  authorId: number;
}